+++ models/post.go
@@ -8,20 +8,20 @@ import (
 )
 
 type Post struct {
-	ID        int       `json:"id"`
+	ID        int64     `json:"id"`
 	Content   string    `json:"content" validate:"required,max=140"`
-	UserID    int       `json:"user_id" validate:"required"`
+	UserID    int64     `json:"user_id" validate:"required"`
 	User      *User     `json:"user,omitempty" validate:"-"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
 }
 
-func ShowPost(id int) (post Post, err error) {
+func ShowPost(id int64) (post Post, err error) {
 	err = db.First(&post, "id=?", id).Error
 	return post, err
 }
 
-func CreatePost(userID int, content string) (err error) {
+func CreatePost(userID int64, content string) (err error) {
 	post := Post{}
 	post.UserID = userID
 	post.Content = content
@@ -34,7 +34,7 @@ func CreatePost(userID int, content string) (err error) {
 	return nil
 }
 
-func UpdatePost(id int, userID int, content string) (err error) {
+func UpdatePost(id int64, userID int64, content string) (err error) {
 	post, err := ShowPost(id)
 	if err != nil {
 		return err
@@ -53,7 +53,7 @@ func UpdatePost(id int, userID int, content string) (err error) {
 	return nil
 }
 
-func DeletePost(id int, userID int) (err error) {
+func DeletePost(id int64, userID int64) (err error) {
 	post, err := ShowPost(id)
 	if err != nil {
 		return err
